Move image out of about-me layout into slide
Previous attempts to resolve path resolution issues on deployment
haven't worked, so this commit aligns the about-me slide with the
talking-vs-coding slide. Now the image is defined within the markdown
fontmatter, which appears to be more compatible with the build.

Still not sure why {url}/assets/me.png gives a 404...